cache the region of minerals/gas, and move the mineral/gas percepts to the UnitsPerceiver (so all units or the mapagent will get those percepts)
