+++ eisbw/src/main/java/eisbw/percepts/perceivers/UnitsPerceiver.java
@@ -1,5 +1,6 @@
 package eisbw.percepts.perceivers;
 
+import java.util.HashMap;
 import java.util.HashSet;
 //import java.util.LinkedList;
 import java.util.List;
@@ -14,9 +15,11 @@ import eisbw.BwapiUtility;
 import eisbw.percepts.AttackingPercept;
 import eisbw.percepts.EnemyPercept;
 import eisbw.percepts.FriendlyPercept;
+import eisbw.percepts.MineralFieldPercept;
 import eisbw.percepts.Percepts;
 import eisbw.percepts.ResourcesPercept;
 import eisbw.percepts.UnderConstructionPercept;
+import eisbw.percepts.VespeneGeyserPercept;
 import eisbw.units.ConditionHandler;
 import jnibwapi.JNIBWAPI;
 import jnibwapi.Player;
@@ -28,12 +31,95 @@ import jnibwapi.types.UnitType.UnitTypes;
  *
  */
 public class UnitsPerceiver extends Perceiver {
+	private final Map<Unit, Integer> regionCache = new HashMap<>();
+
 	/**
 	 * @param api
 	 *            The BWAPI.
 	 */
 	public UnitsPerceiver(JNIBWAPI api) {
 		super(api);
+	}
+
+	@Override
+	public Map<PerceptFilter, Set<Percept>> perceive(Map<PerceptFilter, Set<Percept>> toReturn) {
+		resourcesPercepts(toReturn);
+		unitsPercepts(toReturn);
+		return toReturn;
+	}
+
+	private void resourcesPercepts(Map<PerceptFilter, Set<Percept>> toReturn) {
+		Player self = this.api.getSelf();
+		if (self != null) { // for tests
+			Set<Percept> resourcePercept = new HashSet<>(1);
+			resourcePercept.add(new ResourcesPercept(self.getMinerals(), self.getGas(), self.getSupplyUsed(),
+					self.getSupplyTotal()));
+			toReturn.put(new PerceptFilter(Percepts.RESOURCES, Filter.Type.ON_CHANGE), resourcePercept);
+		}
+		Set<Percept> minerals = new HashSet<>();
+		Set<Percept> geysers = new HashSet<>();
+		for (Unit u : this.api.getNeutralUnits()) {
+			if (BwapiUtility.isValid(u) && u.getType().isMineralField()) {
+				int region = getRegion(u);
+				MineralFieldPercept mineralfield = new MineralFieldPercept(u.getID(), u.getResources(),
+						u.getPosition().getBX(), u.getPosition().getBY(), region);
+				minerals.add(mineralfield);
+			} else if (BwapiUtility.isValid(u) && u.getType().getID() == UnitTypes.Resource_Vespene_Geyser.getID()) {
+				int region = getRegion(u);
+				VespeneGeyserPercept geyser = new VespeneGeyserPercept(u.getID(), u.getResources(),
+						u.getPosition().getBX(), u.getPosition().getBY(), region);
+				geysers.add(geyser);
+			}
+		}
+		for (Unit u : this.api.getMyUnits()) {
+			if (BwapiUtility.isValid(u) && u.getType().isRefinery()) {
+				int region = getRegion(u);
+				VespeneGeyserPercept geyser = new VespeneGeyserPercept(u.getID(), u.getResources(),
+						u.getPosition().getBX(), u.getPosition().getBY(), region);
+				geysers.add(geyser);
+
+			}
+		}
+		toReturn.put(new PerceptFilter(Percepts.MINERALFIELD, Filter.Type.ALWAYS), minerals);
+		toReturn.put(new PerceptFilter(Percepts.VESPENEGEYSER, Filter.Type.ALWAYS), geysers);
+	}
+
+	private void unitsPercepts(Map<PerceptFilter, Set<Percept>> toReturn) {
+		Set<Percept> newunitpercepts = new HashSet<>();
+		Set<Percept> friendlypercepts = new HashSet<>();
+		Set<Percept> enemypercepts = new HashSet<>();
+		Set<Percept> attackingpercepts = new HashSet<>();
+
+		// perceive friendly units
+		setUnitPercepts(this.api.getMyUnits(), newunitpercepts, friendlypercepts, attackingpercepts);
+		// perceive enemy units
+		setUnitPercepts(this.api.getEnemyUnits(), null, enemypercepts, attackingpercepts);
+
+		if (!friendlypercepts.isEmpty()) {
+			toReturn.put(new PerceptFilter(Percepts.FRIENDLY, Filter.Type.ALWAYS), friendlypercepts);
+		}
+		if (!enemypercepts.isEmpty()) {
+			toReturn.put(new PerceptFilter(Percepts.ENEMY, Filter.Type.ALWAYS), enemypercepts);
+		}
+		if (!attackingpercepts.isEmpty()) {
+			toReturn.put(new PerceptFilter(Percepts.ATTACKING, Filter.Type.ALWAYS), attackingpercepts);
+		}
+		if (!newunitpercepts.isEmpty()) {
+			toReturn.put(new PerceptFilter(Percepts.UNDERCONSTRUCTION, Filter.Type.ALWAYS), newunitpercepts);
+		}
+	}
+
+	/**
+	 * @param u
+	 * @return The region for the given unit (from a cache if it was seen before).
+	 */
+	private int getRegion(Unit u) {
+		Integer region = this.regionCache.get(u);
+		if (region == null) {
+			region = BwapiUtility.getRegion(u, this.api.getMap());
+			this.regionCache.put(u, region);
+		}
+		return region.intValue();
 	}
 
 	/**
@@ -82,41 +168,5 @@ public class UnitsPerceiver extends Perceiver {
 				}
 			}
 		}
-	}
-
-	@Override
-	public Map<PerceptFilter, Set<Percept>> perceive(Map<PerceptFilter, Set<Percept>> toReturn) {
-		Set<Percept> newunitpercepts = new HashSet<>();
-		Set<Percept> friendlypercepts = new HashSet<>();
-		Set<Percept> enemypercepts = new HashSet<>();
-		Set<Percept> attackingpercepts = new HashSet<>();
-
-		// perceive friendly units
-		setUnitPercepts(this.api.getMyUnits(), newunitpercepts, friendlypercepts, attackingpercepts);
-		// perceive enemy units
-		setUnitPercepts(this.api.getEnemyUnits(), null, enemypercepts, attackingpercepts);
-
-		if (!friendlypercepts.isEmpty()) {
-			toReturn.put(new PerceptFilter(Percepts.FRIENDLY, Filter.Type.ALWAYS), friendlypercepts);
-		}
-		if (!enemypercepts.isEmpty()) {
-			toReturn.put(new PerceptFilter(Percepts.ENEMY, Filter.Type.ALWAYS), enemypercepts);
-		}
-		if (!attackingpercepts.isEmpty()) {
-			toReturn.put(new PerceptFilter(Percepts.ATTACKING, Filter.Type.ALWAYS), attackingpercepts);
-		}
-		if (!newunitpercepts.isEmpty()) {
-			toReturn.put(new PerceptFilter(Percepts.UNDERCONSTRUCTION, Filter.Type.ALWAYS), newunitpercepts);
-		}
-
-		Player self = this.api.getSelf();
-		if (self != null) { // for tests
-			Set<Percept> resourcePercept = new HashSet<>(1);
-			resourcePercept.add(new ResourcesPercept(self.getMinerals(), self.getGas(), self.getSupplyUsed(),
-					self.getSupplyTotal()));
-			toReturn.put(new PerceptFilter(Percepts.RESOURCES, Filter.Type.ON_CHANGE), resourcePercept);
-		}
-
-		return toReturn;
 	}
 }
