+++ eisbw/src/main/java/eisbw/units/StarcraftUnitFactory.java
@@ -1,12 +1,11 @@
 package eisbw.units;
 
-import java.util.LinkedList;
+import java.util.ArrayList;
 import java.util.List;
 
 import eisbw.percepts.perceivers.BuildingPerceiver;
 import eisbw.percepts.perceivers.GenericUnitPerceiver;
 import eisbw.percepts.perceivers.IPerceiver;
-import eisbw.percepts.perceivers.WorkerPerceiver;
 import jnibwapi.JNIBWAPI;
 import jnibwapi.Unit;
 
@@ -27,9 +26,6 @@ public class StarcraftUnitFactory {
 		this.api = api;
 	}
 
-	// These perceptgenerators are only added on init, so a building that can
-	// fly can not move when built, so please take caution and think when adding
-	// percepts
 	/**
 	 * Creates a unit.
 	 *
@@ -38,16 +34,11 @@ public class StarcraftUnitFactory {
 	 * @return - a StarCraft unit with perceivers.
 	 */
 	public StarcraftUnit create(Unit unit) {
-		List<IPerceiver> perceptGenerators = new LinkedList<>();
+		List<IPerceiver> perceptGenerators = new ArrayList<>(2);
 		perceptGenerators.add(new GenericUnitPerceiver(this.api, unit));
-
 		if (unit.getType().isBuilding()) {
 			perceptGenerators.add(new BuildingPerceiver(this.api, unit));
 		}
-		if (unit.getType().isWorker()) {
-			perceptGenerators.add(new WorkerPerceiver(this.api, unit));
-		}
-
 		return new StarcraftUnit(perceptGenerators, unit.getType().isWorker());
 	}
 }
